Rename topstories package to presenter and inject Scheduler to presenter
